refactor(upload): extract image upload section from UploadTemplate into standalone ImageUploadField component
- Moved image upload logic and UI from `UploadTemplate` to new `ImageUploadField` component
- Simplified `UploadTemplate` structure for better readability and reusability
- Preserved drag-and-drop and validation behaviors within `ImageUploadField`
